requirements.txt
dvc==3.43.1
selenium==4.17.2
tqdm==4.66.1
dvc_s3==3.0.1
yt-dlp==2023.12.30
webvtt-py==0.4.6
langchain==0.1.7
pymilvus==2.3.6
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-slugify==8.0.4
